hopefully refactor everything to be ready
