+++ .vscode/launch.json
@@ -19,7 +19,7 @@
 				"${workspaceFolder}/out/**/*.js",
 				"!${workspaceFolder}/out/web/**"
 			],
-			"preLaunchTask": "npm: watch"
+			"preLaunchTask": "npm: watch:desktop"
 		},
 		{
 			"name": "Run Web Extension",
@@ -33,7 +33,7 @@
 			"outFiles": [
 				"${workspaceFolder}/out/web/**/*.js"
 			],
-			"preLaunchTask": "npm: watch-web"
+			"preLaunchTask": "npm: watch:web"
 		}
 	]
 }
